Updated Text Box + Changelog
Updated Textbox and made it reshow after a version update
