+++ lang/en.json
@@ -14,12 +14,14 @@
             "hint": "Indicates whether the dialog has been shown for this user",
             "title": "Elkan 5e",
             "content": {
-                "header": "Hope you're enjoying the Elkan 5e module!",
+                "header": "Hope you're enjoying the Elkan 5e module",
                 "updateNotice": "Our latest update has included several breaking changes. <b>If you were using our content in DND5e v3.3 and have just updated to use DND5e v4+ </b>, you will find many of your previous characters, NPCs, and features may not be working properly. We recommend you re-create all player characters and replace features on NPCs as needed.",
-                "joinUs": "Join us on <a href=\"https://discord.gg/DNdtcJYGeB\" target=\"_blank\">Discord</a><br>Visit our <a href=\"https://www.elkan5e.com\" target=\"_blank\">Website</a>",
+                "joinUs": "Join us on <a href=\"https://discord.gg/DNdtcJYGeB\" target=\"_blank\">Discord</a> or <a href=\"https://www.elkan5e.com\" target=\"_blank\">Visit our Website</a>",
+                "joinUsText": "For the latest Elkan 5e news or details on upcoming events!",
                 "supportUs": "Support Us",
                 "supportText": "You can support us on <b><a href=\"https://www.patreon.com/Elkan5e\" target=\"_blank\">Patreon</a>.</b> <br> If you would prefer to make a one-time donation, you can donate on <b><a href=\"https://ko-fi.com/elkan5e\" target=\"_blank\">Ko-fi</a></b>."
-            }
+            },
+            "doNotShow": "Do Not Show Again"
         },
         "conditions": {
             "name": "Use Custom Conditions",
