+++ src/module/initalizing.mjs
@@ -6,7 +6,6 @@ import { references } from "./rules/references.mjs";
 import { tools } from "./rules/tools.mjs";
 import { weapons } from "./rules/weapon.mjs";
 import { scroll } from "./rules/scroll.mjs";
-import { gameSettingRegister } from "./gameSettings/gameSettingRegister.mjs";
 
 /**
  * Initializes all Elkan 5e modules in Foundry.
